perf(core): Make `Profile` an immutable fixed size collection of ballots
Profiles are not meant to be growable, but represent some immutable collection of ballots. Therefore they are semantically a Box<B: Ballot>, and that is precisely what this commit does.

This raises the question of why this commit is classified as perf and not refactor. This is because Vec<B: Ballot> is still fine, just not the best choice. Having a Box<B: Ballot> is semantically better, but its primarily been implemented for performance reasons and not semantic ones.