perf(core)!: Replace `Strategy` with its definition inside `VotingBlock`
Having a strategy struct but not a preference struct, method struct, etc is a odd annoyance. Inside VotingBlock, I replaced strategy with its inner implementation.

This change causes downstream effects. In order to easily make a voting block, there now needs to be a builder, and with that, the main election api has to add a `VotingBlock` as an object, instead of its individual components. Thus a builder no longer makes sense, as that was the main point of the builder structure. Now it makes sense to revert back to passing in the structs directly.

BREAKING CHANGE: The API has reverted to passing in the conditions, candidates, voting pools, and method directly with Election::new(...).