fix(mallows): fix panic in Mallows preference model
For |C| > 3, there was a panic because when calculating `min_value`, `remaining_inv` was smaller than `((slots) * (slots - 1) / 2))` which caused the usize to underflow. This meant that `rng.random_range(min_value..=max_value)` was empty, and thus the panic.

Now if this is the case, we simply select 0 since we know that there are enough slots for 0 inversions.