perf(election)!: confirm that candidate id's are small
Now we validate that the candidate Id's are unique and range from 0 to n-1 (where n = # of candidates). This allows us to index the candidates with a vector, where the index i corresponds to the candidate with that Id. The ultimate goal is to remove hashmaps later for performance reasons.

BREAKING CHANGE: Constructing an election now returns a result that needs to be handled.